requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

pydantic>=2.7.0
python-multipart>=0.0.9

openai>=1.40.0
tiktoken>=0.7.0

chromadb>=0.5.3

pdfplumber>=0.11.4
python-docx>=1.1.2
beautifulsoup4>=4.12.3

loguru>=0.7.2
python-dotenv>=1.0.1
